requirements.txt
# Django Core
Django==6.0
asgiref==3.11.0
sqlparse==0.5.5

# Django REST Framework
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1

# Django Extensions
django-cors-headers==4.9.0

# Dependencies
PyJWT==2.10.1
httpx==0.28.1
